Fix uintptr_t size

+++ src/lib/common.h
@@ -7,7 +7,7 @@ typedef int uint32_t;
 typedef long long uint64_t;
 typedef long double uint128_t;
 
-typedef uint32_t uintptr_t;
+typedef uint64_t uintptr_t;
 
 #define NULL 0
